Added/Modified layouts
(1) Added a sidebars.handlebars as we need that on all pages after login.
(2) Altered the layout of the settings.handlebars and CSS so everything was on one vertical line.
(3) Got rid of the sidebars class in homepage handlebars since we have a template for it.

Had to be manually copied from keona/handlebars-css